+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.19.0] - 2020-05-27
+
 ### Changed
 
 - Using `SkeletonPiece` component from `vtex.my-account-commons`.